fix bug

+++ backend/database/migrations/2019_10_30_153149_create_solutions_table.php
@@ -14,7 +14,7 @@ class CreateSolutionsTable extends Migration {
         Schema::create('solutions', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('text');
-            $table->integer('question_id')->unsigned();
+            $table->bigInteger('question_id')->unsigned();
             $table->foreign('question_id')
                   ->references('id')->on('questions')
                   ->onDelete('cascade');
